+++ .gitignore
@@ -97,6 +97,8 @@ docs/_build/
 
 # Pycharm
 .idea/
+/venv/
+
 
 ############
 ## Windows
