+++ hedvalidation/tests/test_warning_reporter.py
@@ -1,5 +1,5 @@
 import unittest
-from hedvalidation import warning_reporter
+from hedvalidator import warning_reporter
 
 
 class Test(unittest.TestCase):
